+++ src/preload-api.js
@@ -15,31 +15,4 @@ const CHANNELS = Object.freeze({
   updateSettings: 'qq-local-recall:update-settings',
 });
 
-function createPreloadApi(
-  ipcRenderer,
-  { includeOpen = false, includeRecovered = false, includeSettings = false } = {},
-) {
-  const api = {
-    listConversations: () => ipcRenderer.invoke(CHANNELS.list),
-    listRecords: peerKey => ipcRenderer.invoke(CHANNELS.listRecords, peerKey),
-    deleteConversations: peerKeys => ipcRenderer.invoke(CHANNELS.delete, peerKeys),
-    deleteRecord: (peerKey, msgId) => ipcRenderer.invoke(CHANNELS.deleteRecord, peerKey, msgId),
-    onRecordsDeleted: callback => {
-      ipcRenderer.on(CHANNELS.deleted, (_event, payload) => callback(payload));
-    },
-  };
-  if (includeOpen) api.openManager = () => ipcRenderer.invoke(CHANNELS.open);
-  if (includeRecovered) {
-    api.persistRenderedMedia = value => ipcRenderer.invoke(CHANNELS.persistMedia, value);
-    api.onRecovered = callback => {
-      ipcRenderer.on(CHANNELS.recovered, (_event, payload) => callback(payload));
-    };
-  }
-  if (includeSettings) {
-    api.getSettings = () => ipcRenderer.invoke(CHANNELS.settings);
-    api.updateSettings = value => ipcRenderer.invoke(CHANNELS.updateSettings, value);
-  }
-  return Object.freeze(api);
-}
-
-module.exports = { CHANNELS, createPreloadApi };
+module.exports = { CHANNELS };