+++ src/preload.js
@@ -3,8 +3,6 @@
 const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('qqLocalRecall', Object.freeze({
-  listConversations: () => ipcRenderer.invoke('qq-local-recall:list-conversations'),
-  deleteConversations: peerKeys => ipcRenderer.invoke('qq-local-recall:delete-conversations', peerKeys),
   openManager: () => ipcRenderer.invoke('qq-local-recall:open-manager'),
   persistRenderedMedia: value => ipcRenderer.invoke('qq-local-recall:persist-rendered-media', value),
   onRecordsDeleted: callback => {